consolidate tests under package
