+++ prompthelix/TODO.md
@@ -5,4 +5,4 @@ General improvements for the `prompthelix` package.
 - [ ] Split modules into smaller packages where appropriate
 - [ ] Enforce consistent logging across modules
 - [ ] Provide configuration for enabling/disabling debug output
-- [x] Clean up placeholder scripts in `tests/interactive` and `tests/oauth2`
+- [x] Clean up placeholder scripts in `tests/interactive` and `tests/oauth2` (tests now consolidated under `prompthelix/tests/`)
