--- tests/interactive/__init__.py
@@ -1,6 +1,0 @@
-"""Interactive test helpers."""
-
-from prompthelix.tests.interactive.base import InteractiveTestCase
-
-__all__ = ["InteractiveTestCase"]
-
