--- tests/oauth2/__init__.py
@@ -1,1 +1,0 @@
-# This space intentionally left blank.
